Group into Executor, Compiler, and IR
