+++ compiler/instruction/expressions/builtins/binary.rs
@@ -4,61 +4,65 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-use std::marker::PhantomData;
+use std::{marker::PhantomData, ops::Rem};
 
 use encoding::value::{value::DBValue, value_type::ValueTypeCategory};
 
-use crate::{
+use crate::instruction::{
     expressions::{
         evaluator::ExpressionEvaluationState,
         expression_compiler::{ExpressionInstruction, ExpressionTreeCompiler, SelfCompiling},
         op_codes::ExpressionOpCode,
         ExpressionEvaluationError,
     },
-    ExpressionCompilationError,
 };
 
-pub trait UnaryExpression<T1: DBValue, R: DBValue> {
+pub trait BinaryExpression<T1: DBValue, T2: DBValue, R: DBValue> {
     const OP_CODE: ExpressionOpCode;
-    fn evaluate(a1: T1) -> Result<R, ExpressionEvaluationError>;
+    fn evaluate(a1: T1, a2: T2) -> Result<R, ExpressionEvaluationError>;
 }
 
-pub struct Unary<T1, R, F>
+pub struct Binary<T1, T2, R, F>
 where
     T1: DBValue,
+    T2: DBValue,
     R: DBValue,
-    F: UnaryExpression<T1, R>,
+    F: BinaryExpression<T1, T2, R>,
 {
-    phantom: PhantomData<(T1, R, F)>,
+    pub phantom: PhantomData<(T1, T2, R, F)>,
 }
 
-impl<T1, R, F> ExpressionInstruction for Unary<T1, R, F>
+impl<T1, T2, R, F> ExpressionInstruction for Binary<T1, T2, R, F>
 where
     T1: DBValue,
+    T2: DBValue,
     R: DBValue,
-    F: UnaryExpression<T1, R>,
+    F: BinaryExpression<T1, T2, R>,
 {
     const OP_CODE: ExpressionOpCode = F::OP_CODE;
     fn evaluate<'a>(state: &mut ExpressionEvaluationState<'a>) -> Result<(), ExpressionEvaluationError> {
+        let a2: T2 = T2::form_db_value(state.pop_value()).unwrap();
         let a1: T1 = T1::form_db_value(state.pop_value()).unwrap();
-        state.push_value(F::evaluate(a1)?.to_db_value());
+        state.push_value(F::evaluate(a1, a2)?.to_db_value());
         Ok(())
     }
 }
 
-impl<T1, R, F> SelfCompiling for Unary<T1, R, F>
+impl<T1, T2, R, F> SelfCompiling for Binary<T1, T2, R, F>
 where
     T1: DBValue,
+    T2: DBValue,
     R: DBValue,
-    F: UnaryExpression<T1, R>,
+    F: BinaryExpression<T1, T2, R>,
 {
     fn return_value_category(&self) -> Option<ValueTypeCategory> {
         Some(R::VALUE_TYPE_CATEGORY)
     }
 
     fn validate_and_append(builder: &mut ExpressionTreeCompiler<'_>) -> Result<(), ExpressionCompilationError> {
+        let a2 = builder.pop_type_single()?;
         let a1 = builder.pop_type_single()?;
-        if a1 != T1::VALUE_TYPE_CATEGORY {
+        if (a1, a2) != (T1::VALUE_TYPE_CATEGORY, T2::VALUE_TYPE_CATEGORY) {
             Err(ExpressionCompilationError::InternalUnexpectedValueType)?;
         }
         builder.push_type_single(R::VALUE_TYPE_CATEGORY);
@@ -67,24 +71,21 @@ where
     }
 }
 
-macro_rules! unary_instruction {
-    ( $( $name:ident = $impl_name:ident($a1:ident: $t1:ty) -> $r:ty $impl_code:block )* ) => { $(
-        pub type $name = Unary<$t1, $r, $impl_name>;
+macro_rules! binary_instruction {
+    ( $( $name:ident = $impl_name:ident($a1:ident: $t1:ty, $a2:ident: $t2:ty) -> $r:ty $impl_code:block )* ) => { $(
+        pub type $name = Binary<$t1, $t2, $r, $impl_name>;
         pub struct $impl_name {}
-        impl UnaryExpression<$t1, $r> for $impl_name {
+        impl BinaryExpression<$t1, $t2, $r> for $impl_name {
             const OP_CODE: ExpressionOpCode = ExpressionOpCode::$name;
-            fn evaluate($a1: $t1) -> Result<$r, ExpressionEvaluationError> {
+            fn evaluate($a1: $t1, $a2: $t2) -> Result<$r, ExpressionEvaluationError> {
                 $impl_code
             }
         })*
     };
 }
 
-pub(crate) use unary_instruction;
-unary_instruction! {
-    MathAbsLong = MathAbsLongImpl(a1: i64) -> i64 { Ok(i64::abs(a1)) }
-    MathAbsDouble = MathAbsDoubleImpl(a1: f64) -> f64 { Ok(f64::abs(a1)) }
-    MathRoundDouble = MathRoundDoubleImpl(a1: f64) -> i64 { Ok(f64::round_ties_even(a1) as i64) } // TODO: Should this be round_ties_even?
-    MathCeilDouble = MathCeilDoubleImpl(a1: f64) -> i64 { Ok(f64::ceil(a1) as i64) }
-    MathFloorDouble = MathFloorDoubleImpl(a1: f64) -> i64 { Ok(f64::floor(a1) as i64) }
+pub(crate) use binary_instruction;
+use crate::inference::ExpressionCompilationError;
+binary_instruction! {
+    MathRemainderLong = MathRemainderLongImpl(a1: i64, a2: i64) -> i64 { Ok(i64::rem(a1, a2)) }
 }
