+++ compiler/planner/program_plan.rs
@@ -4,17 +4,23 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
 use std::collections::HashMap;
 
 use encoding::graph::definition::definition_key::DefinitionKey;
-use ir::inference::type_inference::TypeAnnotations;
-
-use crate::planner::{function_plan::FunctionPlan, pattern_plan::PatternPlan};
+use crate::inference::type_annotations::TypeAnnotations;
+use crate::planner::function_plan::FunctionPlan;
+use crate::planner::pattern_plan::PatternPlan;
 
 pub struct ProgramPlan {
-    pub(crate) entry: PatternPlan,
-    pub(crate) entry_annotations: TypeAnnotations,
-    pub(crate) functions: HashMap<DefinitionKey<'static>, FunctionPlan>,
+    pub entry: PatternPlan,
+    pub entry_annotations: TypeAnnotations,
+    pub functions: HashMap<DefinitionKey<'static>, FunctionPlan>,
 }
 
 impl ProgramPlan {
