+++ executor/instruction/has_executor.rs
@@ -27,7 +27,7 @@ use lending_iterator::{
 };
 use storage::{key_range::KeyRange, snapshot::ReadableSnapshot};
 
-use crate::executor::{
+use crate::{
     batch::ImmutableRow,
     instruction::{
         iterator::{inverted_instances_cache, SortedTupleIterator, TupleIterator},
