+++ executor/instruction/tuple.rs
@@ -11,7 +11,7 @@ use concept::{
 };
 use lending_iterator::higher_order::Hkt;
 
-use crate::executor::VariablePosition;
+use crate::VariablePosition;
 
 #[derive(Debug, Clone)]
 pub(crate) enum Tuple<'a> {
