+++ executor/tests/executor_has.rs
@@ -5,6 +5,8 @@
  */
 
 use std::{borrow::Cow, collections::HashMap, sync::Arc};
+use compiler::inference::annotated_functions::CompiledSchemaFunctions;
+use compiler::inference::type_inference::infer_types;
 
 use concept::{
     error::ConceptReadError,
@@ -13,11 +15,10 @@ use concept::{
 };
 use encoding::value::{label::Label, value::Value, value_type::ValueType};
 use ir::{
-    inference::type_inference::infer_types,
     pattern::constraint::IsaKind,
     program::{
         block::FunctionalBlock,
-        program::{CompiledSchemaFunctions, Program},
+        program::{Program},
     },
 };
 use lending_iterator::LendingIterator;
@@ -26,12 +27,10 @@ use storage::{
     snapshot::{CommittableSnapshot, ReadSnapshot, WriteSnapshot},
     MVCCStorage,
 };
-use traversal::{
-    executor::{batch::ImmutableRow, program_executor::ProgramExecutor},
-    planner::{
-        pattern_plan::{Instruction, IntersectionStep, IterateBounds, PatternPlan, Step},
-        program_plan::ProgramPlan,
-    },
+use executor::{batch::ImmutableRow, program_executor::ProgramExecutor};
+use compiler::planner::{
+    pattern_plan::{Instruction, IntersectionStep, IterateBounds, PatternPlan, Step},
+    program_plan::ProgramPlan,
 };
 
 use crate::common::{load_managers, setup_storage};
@@ -167,7 +166,7 @@ fn traverse_has_unbounded_sorted_from() {
     let executor = {
         let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
         let (_, thing_manager) = load_managers(storage.clone());
-        ProgramExecutor::new(program_plan, &snapshot, &thing_manager).unwrap()
+        ProgramExecutor::new(&program_plan, &snapshot, &thing_manager).unwrap()
     };
 
     {
@@ -232,7 +231,7 @@ fn traverse_has_unbounded_sorted_to_merged() {
     let executor = {
         let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
         let (_, thing_manager) = load_managers(storage.clone());
-        ProgramExecutor::new(program_plan, &snapshot, &thing_manager).unwrap()
+        ProgramExecutor::new(&program_plan, &snapshot, &thing_manager).unwrap()
     };
 
     {
@@ -319,7 +318,7 @@ fn traverse_has_reverse_unbounded_sorted_from() {
     let executor = {
         let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
         let (_, thing_manager) = load_managers(storage.clone());
-        ProgramExecutor::new(program_plan, &snapshot, &thing_manager).unwrap()
+        ProgramExecutor::new(&program_plan, &snapshot, &thing_manager).unwrap()
     };
 
     {
