+++ executor/tests/executor_rp.rs
@@ -5,6 +5,10 @@
  */
 
 use std::{borrow::Cow, collections::HashMap, sync::Arc};
+use compiler::inference::annotated_functions::CompiledSchemaFunctions;
+use compiler::inference::type_inference::infer_types;
+use compiler::planner::pattern_plan::{Instruction, IntersectionStep, IterateBounds, PatternPlan, Step};
+use compiler::planner::program_plan::ProgramPlan;
 
 use concept::{
     error::ConceptReadError,
@@ -12,27 +16,19 @@ use concept::{
     type_::{Ordering, OwnerAPI, PlayerAPI},
 };
 use encoding::value::{label::Label, value::Value, value_type::ValueType};
-use ir::{
-    inference::type_inference::infer_types,
-    pattern::constraint::IsaKind,
-    program::{
-        block::FunctionalBlock,
-        program::{CompiledSchemaFunctions, Program},
-    },
-};
-use lending_iterator::LendingIterator;
+use ir::pattern::constraint::IsaKind;
+use ir::program::block::FunctionalBlock;
+use ir::program::program::Program;
+
 use storage::{
     durability_client::WALClient,
     snapshot::{CommittableSnapshot, ReadSnapshot, WriteSnapshot},
     MVCCStorage,
 };
-use traversal::{
-    executor::{batch::ImmutableRow, program_executor::ProgramExecutor},
-    planner::{
-        pattern_plan::{Instruction, IntersectionStep, IterateBounds, PatternPlan, Step},
-        program_plan::ProgramPlan,
-    },
-};
+use executor::batch::ImmutableRow;
+use executor::program_executor::ProgramExecutor;
+use lending_iterator::LendingIterator;
+
 
 use crate::common::{load_managers, setup_storage};
 
@@ -206,7 +202,7 @@ fn traverse_rp_unbounded_sorted_from() {
     let executor = {
         let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
         let (_, thing_manager) = load_managers(storage.clone());
-        ProgramExecutor::new(program_plan, &snapshot, &thing_manager).unwrap()
+        ProgramExecutor::new(&program_plan, &snapshot, &thing_manager).unwrap()
     };
 
     {
@@ -272,7 +268,7 @@ fn traverse_has_unbounded_sorted_to_merged() {
     let executor = {
         let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
         let (_, thing_manager) = load_managers(storage.clone());
-        ProgramExecutor::new(program_plan, &snapshot, &thing_manager).unwrap()
+        ProgramExecutor::new(&program_plan, &snapshot, &thing_manager).unwrap()
     };
 
     {
