+++ function/lib.rs
@@ -5,10 +5,10 @@
  */
 
 use std::{error::Error, fmt};
+use compiler::inference::TypeInferenceError;
 
 use encoding::error::EncodingError;
 use ir::{
-    inference::TypeInferenceError,
     program::{FunctionDefinitionError, FunctionReadError},
 };
 use storage::{snapshot::SnapshotGetError, ReadSnapshotOpenError};
