--- inference/BUILD
@@ -1,37 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-load("@vaticle_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
-load("@rules_rust//rust:defs.bzl", "rust_library")
-package(default_visibility = ["//visibility:public"])
-
-rust_library(
-    name = "inference",
-    crate_root = "lib.rs",
-    srcs = glob([
-        "*.rs",
-        "expressions/*.rs",
-        "expressions/builtins/*.rs",
-    ]),
-    deps = [
-        "//answer",
-        "//concept",
-        "//encoding",
-        "//ir",
-        "//storage",
-
-        "@vaticle_typeql//rust:typeql",
-        "@crates//:itertools",
-        "@crates//:chrono",
-    ],
-)
-
-checkstyle_test(
-    name = "checkstyle",
-    include = glob(["*", "*/*.rs", "expressions/builtins/*.rs"]),
-    exclude = glob([
-        "Cargo.*",
-    ]),
-    license_type = "mpl-header",
-)
